+++ NEWS.md
@@ -1,4 +1,4 @@
-# mmrm 0.3.7.9005
+# mmrm 0.3.7.9006
 
 ### New Features